+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - [#115](https://github.com/crypto-com/pystarport/pull/115) avoid cli redundant migrated key log in stdout.
 - [#117](https://github.com/crypto-com/pystarport/pull/117) make event_query_tx_for optional.
 - [#121](https://github.com/crypto-com/pystarport/pull/121) Support sdk 0.50
+- [#122](https://github.com/crypto-com/pystarport/pull/122) Adjust status and staking related cli for sdk 0.50
 
 
 *Feb 7, 2023*